Shortened names of the files & removed install.txt

+++ resetusb.sh
@@ -6,12 +6,18 @@
 # Run the script with sudo if your printer does not print after first job.
 
 bus_id=$(lsusb | grep Ricoh | cut -d" " -f2)
+
+if [[ -z $bus_id ]]; then
+  printf "No Ricoh printer found. Make sure it's connected to the computer and turned on.\\n"
+  exit 1
+fi
+
 dev_id=$(lsusb | grep Ricoh | cut -d" " -f4 | cut -c1-3)
 
-exec_name="reset_usb"
-cc reset_usb.c -o "$exec_name"
+exec_name="resetusb"
+cc resetusb.c -o "$exec_name"
 
-./"$exec_name" "/dev/bus/usb/$bus_id/$dev_id"
+"./$exec_name" "/dev/bus/usb/$bus_id/$dev_id"
 
 rm -f "$exec_name"
